JAMES-2712 Extract ExportRequest object
